:fire: removed old trainboard project and renamed new one to TrainBoard from TrainBoard1
