--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "OpenLDBWS"]
-	path = OpenLDBWS
-	url = https://github.com/hazzaRR/OpenLDBWS.git
