+++ TrainBoard/Entities/ScreenData.cs
