+++ TrainBoard/Workers/DisplayWorker.cs
