--- TrainBoard1/Program.cs
@@ -1,15 +1,0 @@
-using TrainBoard.Services;
-using TrainBoard.Workers;
-
-var builder = Host.CreateApplicationBuilder(args);
-
-builder.Services.AddSingleton<IRgbMatrixService, RgbMatrixService>(); 
-builder.Services.AddHostedService<DataFeedWorker>();
-builder.Services.AddHostedService<DisplayWorker>();
-// builder.Services.AddHostedService<CallingPointsWorker>();
-// builder.Services.AddHostedService<TogglePlatformAndEtdWorker>();
-// builder.Services.AddHostedService<CurrentTimeWorker>();
-builder.Services.AddMemoryCache();
-
-var host = builder.Build();
-host.Run();
