--- TrainBoard1/TrainBoard.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Worker">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <UserSecretsId>dotnet-TrainBoard-6445457c-9d31-4266-9f45-ab71848cf1d7</UserSecretsId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
-    <ProjectReference Include="../RPiRgbLEDMatrix/RPiRgbLEDMatrix.csproj" />
-    <ProjectReference Include="../OpenLDBWS/OpenLDBWS.csproj" />
-    <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="9.0.1-*" />
-  </ItemGroup>
-</Project>
